move useOrganizationQuery to separate file
also move related interface and enum
