+++ jsapp/js/account/billingContextProvider.component.tsx
@@ -3,7 +3,7 @@ import { OneTimeAddOnsContext, useOneTimeAddOns } from './useOneTimeAddonList.ho
 import {UsageContext, useUsage} from 'js/account/usage/useUsage.hook';
 import {ProductsContext, useProducts} from 'js/account/useProducts.hook';
 import sessionStore from 'js/stores/session';
-import {useOrganizationQuery} from 'js/account/stripe.api';
+import {useOrganizationQuery} from 'js/account/organization/organizationQuery';
 
 export const BillingContextProvider = (props: {children: ReactNode}) => {
   const orgQuery = useOrganizationQuery();
@@ -11,7 +11,7 @@ export const BillingContextProvider = (props: {children: ReactNode}) => {
   if (!sessionStore.isLoggedIn) {
     return <>{props.children}</>;
   }
-  
+
   const usage = useUsage(orgQuery.data?.id || null);
   const products = useProducts();
   const oneTimeAddOns = useOneTimeAddOns();
