+++ jsapp/js/account/plans/plan.component.tsx
@@ -9,7 +9,7 @@ import React, {
 } from 'react';
 import {useNavigate, useSearchParams} from 'react-router-dom';
 import styles from './plan.module.scss';
-import {postCheckout, postCustomerPortal, useOrganizationQuery} from '../stripe.api';
+import {postCheckout, postCustomerPortal} from '../stripe.api';
 import Button from 'js/components/common/button';
 import classnames from 'classnames';
 import LoadingSpinner from 'js/components/common/loadingSpinner';
@@ -28,7 +28,6 @@ import {
 } from 'js/account/stripe.utils';
 import type {
   Price,
-  Organization,
   Product,
   SubscriptionInfo,
   SinglePricedProduct,
@@ -39,6 +38,7 @@ import {PlanContainer} from 'js/account/plans/planContainer.component';
 import {ProductsContext} from '../useProducts.hook';
 import {ACCOUNT_ROUTES} from 'js/account/routes.constants';
 import {useRefreshApiFetcher} from 'js/hooks/useRefreshApiFetcher.hook';
+import {useOrganizationQuery, type Organization} from 'js/account/organization/organizationQuery';
 
 export interface PlanState {
   subscribedProduct: null | SubscriptionInfo[];
