+++ jsapp/js/account/plans/planButton.component.tsx
@@ -1,8 +1,8 @@
 import BillingButton from 'js/account/plans/billingButton.component';
-import React, {useContext} from 'react';
 import type {Price, SinglePricedProduct} from 'js/account/stripe.types';
-import {postCustomerPortal, useOrganizationQuery} from 'js/account/stripe.api';
+import {postCustomerPortal} from 'js/account/stripe.api';
 import {processCheckoutResponse} from 'js/account/stripe.utils';
+import {useOrganizationQuery} from 'js/account/organization/organizationQuery';
 
 interface PlanButtonProps {
   buySubscription: (price: Price, quantity?: number) => void;
