+++ jsapp/js/router/RequireOrg.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import LoadingSpinner from 'js/components/common/loadingSpinner';
-import {useOrganizationQuery} from 'js/account/stripe.api';
+import {useOrganizationQuery} from 'js/account/organization/organizationQuery';
 
 interface Props {
   children: React.ReactNode;
